consumer/requirements.txt
# logging
kafka-python
kq
pymongo
# requests
# pymongo
# uuid